refactor(layout): change layout folder structure (#47)
