+++ apps/site/src/app/components/index.ts
@@ -1,2 +1,2 @@
 export * from "./app-name";
-export * from "./default-layout";
+export * from "./layout";
