+++ apps/site/src/app/utils/index.ts
@@ -1,2 +1,1 @@
-export * from "./layout";
 export * from "./routing";
