+++ libs/features/feedback/src/components/layout/layout.tsx
@@ -1,4 +1,4 @@
-import { BasicLayout } from "@patronage-web/shared";
+import { Layout as BasicLayout } from "@patronage-web/shared";
 
 import { footerConfig } from "./footer-config";
 import { navbarConfig } from "./navbar-config";
