+++ libs/features/feedback/src/views/home.tsx
@@ -1,13 +1,12 @@
 import { ActionCard, FeedbackRoute } from "@patronage-web/shared";
 import { useTranslation } from "react-i18next";
 
-import { StarsImage } from "../../assets";
+import { StarsImage } from "../assets";
 
 export const HomeView: React.FC = () => {
   const { t } = useTranslation();
 
   return (
-    // TODO - add layout when proper data will be delivered
     <ActionCard
       description={t("homepage.newPresentationDescription")}
       image={StarsImage}
