+++ libs/shared/src/components/index.ts
@@ -3,6 +3,7 @@ export * from "./base-button";
 export * from "./basic-info";
 export * from "./footer";
 export * from "./language-button";
+export * from "./layout";
 export * from "./loader";
 export * from "./localized-link";
 export * from "./navbar";
