+++ libs/shared/src/components/layout/index.tsx
@@ -1,13 +1,13 @@
-import { Footer, FooterProps } from "../../components/footer";
-import { Navbar, NavbarProps } from "../../components/navbar";
+import { Footer, FooterProps } from "../footer";
+import { Navbar, NavbarProps } from "../navbar";
 import * as Styled from "./style";
 
-export interface BasicLayoutProps {
+export interface LayoutProps {
   navbarConfig: NavbarProps["config"];
   footerConfig: FooterProps["config"];
 }
 
-export const BasicLayout: React.FC<BasicLayoutProps> = ({ children, navbarConfig, footerConfig }) => (
+export const Layout: React.FC<LayoutProps> = ({ children, navbarConfig, footerConfig }) => (
   <Styled.LayoutBox>
     <Navbar color='inherit' position='sticky' config={navbarConfig} />
     <Styled.ContentBox>{children}</Styled.ContentBox>
