+++ libs/shared/src/index.ts
@@ -2,7 +2,6 @@ export * from "./assets";
 export * from "./components";
 export * from "./constants";
 export * from "./hooks";
-export * from "./layouts";
 export * from "./types";
 export * from "./utils";
 export * from "./views";
