--- libs/shared/src/layouts/index.ts
@@ -1,1 +1,0 @@
-export * from "./basic-layout";
